Restructure DarkNet to TensorFlow code
